+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add support for extra CTest environment variables. This addresses issue #33. 
+
 ## [0.14.2] - 2021-11-11
 
 ### Fixed